Client (flags-only) and local (flags+scan) subpkgs
This commit adds the two "sub-entrypoints" of the new zgrab2 bin
directory. Files are described inversely to their code dependency:

** bin/client/commands.go **

This imports the zgrab2 modules' structs and generates a set of cobra
CLI commands out of them. Any CLI interaction will thus directly modify
the scans configurations, transparently. This code never knows about the
scanners themselves, and does noot need to.

** bin/default_modules.go **

This file is the only one in the entire zgrab2 project that has access
to both the modules flags and their associated scan code. This is so
that scanners can access their flags and zgrab2 core, while leaving
modules' flags with no dependencies.

This file is imported by either a local zgrab2 binary, or a server-only
binary (no CLI compiled in). However, this file also populates the default
module flags set.

**Additional notes**

- An empty client.go file is here because of some tries, kept here
  either to have some relevant code on the shell, but might be moved
  away soon.
- Comments changes in the root, explaining a function call contents.
- Overall, some packages might disappear on the long run, since we might
  find some efficient ways to solve our current set of constraints.