End of previous commit #d45933b: entries
This commit adds the last file described in the previous commit,
`default_modules.go`, containing the calls for registrating our
scan modules components.

This commit also goes with the explanation of a file that was included
in the previous commit: `bin/local/root.go`, containing many things:

- A root cobra command generated out of our base zgrab2.Config flags
  structure. This command is executed in our cmd/zgrab2 entrypoint.
- An init function in which we fetch ONLY the modules' flags as CLI
  commands, to which we bind the same exec run function.
- An exec run function that simly gets the cobra command that has been
  called, gets the module struct, and passes it directly to our main
  zgrab2 package, since we are here working in a local binary.
- Some old code that is kept as remainder, will be moved soon.

Please refer to the previous commit for the actual file...
and to this file for the previous commit' comment.