Example of the new dependency relationships: TLS
This commit takes a bit of advance by pushing some code only relevant to
explain the new dependency relationships between our components in
zgrab2. This is explained with a basic TLS scanner, which however needs
specialized code to work, thus giving us a good case study.

**Overview of the problems & needs for the TLS scanner**

- The TLS scans needs to open connections with various settings, which
  has pushed zgrab2 devs to write this utility code around the TLS flags
  themselves. This should be preserved, and has been. Nonetheless, this
  utility connection code is not needed by the CLI strictly speaking.
- There was, like other modules, no segregation between flags (and their
  util code) and the scanning code). This is fixed as well.

**Explanations**

- The TLS scanner itself is now in module/tls/scanner.go, and imports
  the TLS flags defined in module/tls.go.
- Meanwhile, zgrab2 core redeclares the module.TLSFlags and adds some
  methods around this type, so as to leverage it for opening custom
  connections.
- The TLS scanner, which has also access to zgrab2 core, casts its
  module.TLSFlags into a zgrab2.TLSFlags, since the memory layout is the
  same, and can therefore use its associated connection code.

**Conclusion**

The TLS scanner is actually one of the most complicated in the sense
that it heavily leverages zgrab2 core, although the scanner itself is
small. This example was brought to show how all modules, when in need of
their flags or of some associated utility code, can import and use them,
while still preserving import of any strictly related CLI code.