All scan modules' flags type definition in module/
This commit contains all flags (modules) type definitions for all
scanners. These modules are defined in module/, which never makes use
of zgrab2 core or scanners code.
All flags go with the methods used to qualify as a "CLI" module
Also contains the module.Scan interface, which used to be
zgrab2.ScanFlags interface